chore(recomp): regenerate the map, and verify the header-aware cache
The cache-invalidation fix from the previous commit is now confirmed by
experiment rather than by reading: appending a line to CM2Lighting.hpp reparses
41 files -- every translation unit that includes it -- and reverting reparses
the same 41. Before the fix it was 0, which is the bug.

The memoised per-file hash holds the cost where it should be: 1s to parse when
nothing changed (12s end to end), 77s for the 41 files a header edit touches,
against the 695s a full rebuild takes.

Also switches the fog colour to the uint32_t GxRsSet overload, matching how
CShaderEffect passes CImVector::value two files over; the explicit int32_t cast
was gratuitous.

Totals unchanged: mapped 3128, faithful 1212, spine 530, render surface 265.

+++ src/model/CM2Lighting.cpp
@@ -318,7 +318,7 @@ void CM2Lighting::SetupGxFog() {
     fogColor.r = CM2Lighting::FogColorByte(this->m_fogColor.x);
     fogColor.a = 0xFF;
 
-    GxRsSet(GxRs_FogColor, static_cast<int32_t>(fogColor.value));
+    GxRsSet(GxRs_FogColor, fogColor.value);
     GxRsSet(GxRs_Fog, 1);
 }
